Added the ambibus script that can be placed in /etc/init.d and used to autostart the script on boot.
Do /etc/init.d/ambibus start to manually start, /etc/init.d/ambibus enable
to start on each boot.

Change the busy wait to a call to 'sleep' so that we don't grab the CPU
while doing nothing.

Alter the display of numbers so that it tries to ensure that there's a
space between the numbers when it can.  This occurs when the left most
number has one digit and the right most has two digits.  In that case
the first two digits are swapped for better readability.